feat: decode packets progressively
this allows to use fields that are already decoded to decode another field (for example in a switch).
also, before this, if a field in the packet had an unknown type, the whole packet's data would be empty. now it returns all the data that was decoded before the field with the unknown type